% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeNeighbors.r
\name{analyzeNeighbors}
\alias{analyzeNeighbors}
\title{Umbrella function that identifies sets of related genome neighborhoods in IMG-derived data}
\usage{
analyzeNeighbors(
  imgGenes = imgGenes,
  imgNeighbors = imgNeighbors,
  efiRepnodes = FALSE,
  neighborThreshold = 0.025,
  geneName = geneName,
  useInterPro = FALSE,
  useHypo = TRUE,
  autoClust = TRUE,
  clustMethod = "tidygraph",
  alphaVal = 0.95,
  bootStrap = 10,
  tgCutoff = 0.6
)
}
\arguments{
\item{imgGenes}{What is the file with the metadata for your gene of interest? Filename as string ("filename.txt")}

\item{imgNeighbors}{What is the file with the metadata for neighbors of your gene of interest? Filename as string ("filename.txt")}

\item{efiRepnodes}{Are values here post-EFI-EST? Boolean, defaults to FALSE.}

\item{neighborThreshold}{In what percentage of neighborhoods must a protein family show up to be of interest?  Number, defaults to 0.025}

\item{geneName}{What is the name of your gene? Gene name as string ("genE")}

\item{useInterPro}{Should InterPro families be used in neighborhood analyses? Boolean, defaults to FALSE}

\item{useHypo}{Should hypothetical protein families from prepNeighbors be used in analyses?  Boolean, defaults to TRUE}

\item{autoClust}{Should clusters be automatically identified? Boolean, defaults to TRUE.}

\item{clustMethod}{What method should be used to identify these clusters? String ("tidygraph" or "pvclust")}

\item{alphaVal}{What alpha value cutoff should be used for pvclust?  Number, defaults to 0.95}

\item{bootStrap}{How many bootstrap rounds for pvclust? Integer, defaults to 10}

\item{tgCutoff}{What sort of edge similarity should be kept for tidygraph? Number (0-1), defaults to 0.65.}
}
\value{
Updated metadata and misc. figures and files en route
}
\description{
This function encompasses several subfunctions that catalog and quantify protein families in the neighborhood of genes of interest and use that data to identify genomic neighborhoods that are themselves similar.
}
\examples{
analyzeNeighborsOutput <- analyzeNeighbors(imgGenes="repnodeGenes.txt", imgNeighbors = "repnodeNeighbors.txt", geneName = "genE") 

}
