% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famInfoList.r
\docType{data}
\name{famInfoList}
\alias{famInfoList}
\title{Basic information for Pfam, TIGRFam, and InterPro families}
\format{
A data frame with three columns (name, shortDesc, and longDesc)
}
\source{
Misc.
}
\usage{
data(famInfoList)
}
\description{
This is a static file containing descriptions of all Pfam, TIGRFam, and InterPro families
Updated 2020.04.09
}
\examples{
data(famInfoList)
}
\references{
None
}
\keyword{datasets}
