% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbToIMG.r
\name{gbToIMG}
\alias{gbToIMG}
\title{Making a fake IMG metadata file from a GenBank file}
\usage{
gbToIMG(
  dataFolder = dataFolder,
  neighborNum = 10,
  goiListInput = goiListInput,
  geneName = "genE",
  removeDupes = TRUE,
  scaffoldGenBase = 3e+09,
  genomeGenBase = 4e+09
)
}
\arguments{
\item{dataFolder}{location of folder containing GenBank files to be processed. Character string.}

\item{neighborNum}{number of neighbors surrounding the gene of interest. Integer.}

\item{goiListInput}{filename of file containing list of locus tags for gene of interest.  Character string.}

\item{geneName}{name of family of genes of interest (for filenaming purposes).  Character string.}

\item{removeDupes}{remove duplicate gene entries if they show up.  Boolean.}
}
\value{

}
\description{
This function takes GenBank files as input and outputs fake IMG-style metadata files for genes of interest and for their neighbors, along with sequences etc.  Replaces generateNeighbors for GenBank input.
}
\examples{
gbToIMGOutput <- gbToIMG(dataFolder="/data/here", neighborNum=5, goiListInput = "goiList.txt", geneName="genE", removeDupes=TRUE)

}
