% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNeighbors.r
\name{generateNeighbors}
\alias{generateNeighbors}
\title{Standalone Function to Generate Lists of Neighboring Genes from IMG gene_oids}
\usage{
generateNeighbors(
  imgGenes = imgGenes,
  neighborNumber = neighborNumber,
  includeGene = TRUE,
  geneName = geneName
)
}
\arguments{
\item{imgGenes}{What is the file with the metadata for your gene of interest? Filename as string ("filename.txt")}

\item{neighborNumber}{How many neighbors do you want to look at on each side of the gene? Integer.}

\item{includeGene}{Do you want your genes of interest in the diagrams? T/F value, defaults to TRUE.}

\item{geneName}{What is the name of your gene? Gene name as string ("genE")}
}
\value{
Lists of neighboring genes to upload to your IMG Gene Cart
}
\description{
This function uses IMG metadatasets to generate relatively attractive gene cluster/genomic neighborhood diagrams that are scaled and that generate vector graphics
}
\examples{
generateNeighborsOutput <- generateNeighbors(imgGenes="geneFile.txt", neighborNumber=10, geneName="genE")

}
