% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifySubgroups.r
\name{identifySubgroups}
\alias{identifySubgroups}
\title{Standalone function to identify subgroups of proteins in larger IMG datasets, based on sequence}
\usage{
identifySubgroups(
  geneList = geneList,
  imgNeighbors = imgNeighbors,
  imgNeighborSeqs = imgNeighborSeqs,
  geneName = geneName,
  subgroupDesc = subgroupDesc,
  cutoffType = cutoffType,
  cutoffValue = cutoffValue,
  sysTerm = sysTerm,
  numThreads = 1,
  alignSubgroups = FALSE,
  defFamNum = 0,
  lightExport = FALSE
)
}
\arguments{
\item{geneList}{File location for list of subset of genes to analyze (single column with gene_oid header). String.}

\item{imgNeighbors}{File location for file with metadata for neighbors of your genes of interest.  String.}

\item{imgNeighborSeqs}{File location for fasta-formatted file with sequence information for neighbors of your gene of interest.  String.}

\item{geneName}{Character string with your gene of interest's name - single word preferred.  Character.}

\item{subgroupDesc}{Character string with the description of gene subset - single word (incl. hyphens) preferred.  Character.}

\item{cutoffType}{The cutoff type being used - "identity" or "evalue".  Character.}

\item{cutoffValue}{The percent ID or evalue cutoff after which edges are not formed when using tidygraph.  Number.}

\item{sysTerm}{The type of terminal (wsl vs linux/unix/macos) from which blastp and other commands will be run.  Character.}

\item{numThreads}{The number of processor threads to be devoted to certain steps. Number, defaults to 1.}

\item{alignSubgroups}{Indicates whether or not MAFFT alignments should be produced for specific subgroups.  Boolean, defaults to FALSE.}

\item{defFamNum}{Starting number for families. Useful when running several consecutive protein types. Number, defaults to 0.}

\item{lightExport}{Indicates whether a simplified export format - gene name and family only - should be used. Boolean, defaults to FALSE.}
}
\value{
Updated metadata for neighboring genes (additional files generated en route)
}
\description{
Given a desired genome neighborhood and lists of genes of interest and their neighbors, along with their protein sequences, this program uses all-by-all-blast & a network-based clustering method to identify subsets of proteins that may represent meaningful subgroups.
}
\examples{
\dontrun{
identifySubgroupsOut <- identifySubgroups(geneList = "geneList.txt", 
                                          imgNeighbors = "imgNeighbors.txt", 
                                          imgNeighborSeqs = "imgneighborSeqs.fa", 
                                          geneName = "genE", 
                                          subgroupDesc="enzymes", 
                                          cutoffType = "identity", 
                                          cutoffValue = 40, 
                                          sysTerm = "nix", 
                                          numThreads = 8)
}

}
