% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorpIprScan.r
\name{incorpIprScan}
\alias{incorpIprScan}
\title{Subfunction to identify subgroups of hypothetical proteins in larger IMG datasets}
\usage{
incorpIprScan(
  iprScanSource = iprScanSource,
  imgNeighborsSource = imgNeighborsSource,
  geneName = geneName,
  addPfam = TRUE,
  addTigrfam = TRUE,
  addIPRfam = TRUE
)
}
\arguments{
\item{iprScanSource}{Data frame from neighborPrep with metadata for your genes of interest. Character string, required.}

\item{imgNeighborsSource}{Data frame from neighborPrep with metadata for neighbors of your genes of interest. Character string, required.}

\item{geneName}{Character string with your gene of interest's name.  Required.}

\item{addPfam}{Should Pfam assignments be added to the metadata table? Boolean, defaults to TRUE.}

\item{addTigrfam}{Should TIGRfam assignments be added to the metadata table? Boolean, defaults to TRUE.}

\item{addIPRfam}{Should InterPro family assignments be added to the metadata table? Boolean, defaults to TRUE.}
}
\value{
Updated metadata for neighboring genes (additional files generated en route)
}
\description{
Given a desired genome neighborhood and lists of genes of interest and their neighbors, along with their protein sequences, this program uses all-by-all-blast & a clustering method of choice to identify subsets of hypothetical proteins that may represent meaningfully associated proteins.
}
\examples{
incorpIprScanOut <- incorpIprScan(iprScanSource = "iprScan.txt", imgNeighborsSource = "fauxNeighborsData.txt", geneName = "genE", addPfam = TRUE, addTigrfam = TRUE)  

}
