% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborCatalog.r
\name{neighborCatalog}
\alias{neighborCatalog}
\title{Subfunction to catalog protein families in IMG metadata file}
\usage{
neighborCatalog(
  imgGenesTrimmed = imgGenesTrimmed,
  imgNeighborsTrimmed = imgNeighborsTrimmed,
  geneName = geneName,
  neighborThreshold = neighborThreshold,
  useInterPro = useInterPro,
  useHypo = useHypo
)
}
\arguments{
\item{imgGenesTrimmed}{Data frame from analyzeNeighbors containing metadata for genes of interest. Character string, required.}

\item{imgNeighborsTrimmed}{Data frame from analyzeNeighbors containing metadata for neighbors of genes of interest. Character string, required.}

\item{geneName}{Name of gene of interest as string. Character string, required.}

\item{neighborThreshold}{Abundance of protein family required to include it in further analysis. Number (0-1)}

\item{useInterPro}{Should InterPro families be used for analysis? Boolean, defaults to FALSE.}

\item{useHypo}{Should hypothetical protein families from prepNeighbors be used for analysis? Boolean, defaults to TRUE.}
}
\value{
List of relevant protein families
}
\description{
This function identifies and quantifies protein families found in genomic neighborhoods (as defined by IMG metadata)
}
\examples{
\dontrun{
neighborCatalogOut <- neighborCatalog(imgGenesTrimmed=imgGenesTrimmed, 
                                      imgNeighborsTrimmed = imgNeighborsTrimmed, 
                                      neighborNumber=neighborNumber, 
                                      includeGene=includeGene, 
                                      geneName=geneName)
}

}
