% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborClusters.r
\name{neighborClusters}
\alias{neighborClusters}
\title{Subfunction to cluster genomic neighborhoods based on the similarity of their protein family sets}
\usage{
neighborClusters(
  imgGenesTrimmed = imgGenesTrimmed,
  imgNeighborsTrimmed = imgNeighborsTrimmed,
  geneName = geneName,
  neighborMatrixData = neighborMatrixData,
  autoClust = autoClust,
  clustMethod = clustMethod,
  alphaVal = alphaVal,
  bootStrap = bootStrap,
  coreGeneName = coreGeneName,
  tgCutoff = tgCutoff
)
}
\arguments{
\item{imgGenesTrimmed}{Data frame from analyzeNeighbors containing metadata for genes of interest. Character string, required.}

\item{imgNeighborsTrimmed}{Data frame from analyzeNeighbors containing metadata for neighbors of genes of interest. Character string, required.}

\item{geneName}{Name of gene of interest as string.  Required.}

\item{neighborMatrixData}{Matrix from analyzeNeighbors/neighborMatrix containing binary data re: protein family presence in gene clusters. Required.}

\item{autoClust}{Should clusters be automatically assigned? Boolean, defaults to TRUE in parent function.}

\item{clustMethod}{String specifying what tool (tidygraph, pvclust) will be used to identify clusters. Defaults to "tidygraph" in parent function.}

\item{alphaVal}{Alpha value cutoff for pvclust. Number, defined in parent function.}

\item{bootStrap}{Bootstrap rounds for pvclust. Integer, defined in parent function.}

\item{coreGeneName}{Name of gene of interest as string, without any suffix. String, defined in parent function.}

\item{tgCutoff}{Number (0-1) representing edge similarity to keep for tidygraph. Default 0.6.}
}
\value{
List with matrix object and updated metadata, with misc. files output along the way
}
\description{
This function cluster-based methods to identify subgroups of gene clusters containing genes of interest that have similar genomic neighborhoods.
}
\examples{
\dontrun{
neighborClustersOut <- neighborClusters(imgGenesTrimmed = imgGenesTrimmed, 
                                        imgNeighborsTrimmed = imgNeighborsTrimmed, 
                                        geneName = "genE", 
                                        neighborMatrixData = neighborMatrixData, 
                                        autoClust = TRUE, 
                                        clustMethod = "tidygraph", 
                                        coreGeneName = "genE", 
                                        tgCutoff=.6)
}

}
