% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborHere.r
\name{neighborHere}
\alias{neighborHere}
\title{Subfunction to quantify protein family presence in the proximity of genes of interest}
\usage{
neighborHere(
  imgNeighborsTrimmed = imgNeighborsTrimmed,
  familyList = familyList,
  geneName = geneName,
  coreGeneName = coreGeneName
)
}
\arguments{
\item{imgNeighborsTrimmed}{Data frame from analyzeNeighbors containing metadata for neighbors of genes of interest}

\item{familyList}{Data from from analyzeNeighbors/neighborCatalog containing a list of protein families for further analysis}

\item{geneName}{Name of gene of interest as string}

\item{coreGeneName}{Name of gene of interest as string, without any suffix.}
}
\value{
Table of binary data indicating presence of absence of a protein family in the proximity of a gene of interest
}
\description{
This subfunction identifies whether members are a given protein family are present in the genomic neighborhood of a gene of interest
}
\examples{
neighborHereOut <- neighborHere(imgNeighborsTrimmed=imgNeighborsTrimmed, familyList = familyList, geneName = geneName, coreGeneName = coreGeneName) 

}
