% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborMatrix.r
\name{neighborMatrix}
\alias{neighborMatrix}
\title{Subfunction to matrix-ify data about protein presence in a given genomic neighborhood}
\usage{
neighborMatrix(
  imgGenesTrimmed = imgGenesTrimmed,
  neighborBinary = neighborBinary,
  familyList = familyList,
  geneName = geneName,
  coreGeneName = coreGeneName
)
}
\arguments{
\item{imgGenesTrimmed}{Data frame from analyzeNeighbors containing metadata for genes of interest}

\item{neighborBinary}{Data frame from analyzeNeighbors/neighborHere binary data regarding presence of protein families in close proximity}

\item{familyList}{Data from analyzeNeighbors/neighborCatalog listing common nearby protein families}

\item{geneName}{Name of gene of interest as string}

\item{coreGeneName}{Name of gene of interest as string, without any suffix.}
}
\value{
Matrix object
}
\description{
This function uses binary present/absent data for various protein families in the proximity of a set of genes of interest and matrixifies it.
}
\examples{
neighborMatrixOut <- neighborMatrix(imgGenesTrimmed=imgGenesTrimmed, neighborBinary = neighborBinary, familyList = familyList, geneName = geneName, coreGeneName = coreGeneName) 

}
