% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborTrim.r
\name{neighborTrim}
\alias{neighborTrim}
\title{Subfunction to clean up IMG metadata files}
\usage{
neighborTrim(
  imgNeighborsData = imgNeighborsData,
  imgGenesData = imgGenesData,
  imgNeighborsContext = imgNeighborsContext,
  trimShortClusters = trimShortClusters,
  neighborNumber = neighborNumber,
  geneName = geneName,
  imgNeighborSeqs = imgNeighborSeqs,
  imgGeneSeqs = imgGeneSeqs
)
}
\arguments{
\item{imgNeighborsData}{Data frame from hypoProteins or neighborPrep with metadata for neighbors of your gene of interest.}

\item{imgNeighborsContext}{Data frame from hypoProteins or neighborPrep tying genes of interest to their neighbors.}

\item{trimShortClusters}{Should gene clusters with fewer than the minimum gene number be visualized? T/F value.}

\item{neighborNumber}{How many neighbors do you want to look at on each side of the gene? Integer from 1 up.}

\item{geneName}{Character string with your gene of interest's name}

\item{imgNeighborSeqs}{Data frame from hypoProteins or neighborPrep with sequence information for neighbors of your gene of interest.}

\item{imgGeneSeqs}{Data frame from hypoProteins or neighborPrep with sequence information for your gene of interest.}

\item{imgGenesTemp}{Data frame from hypoProteins or neighborPrep with metadata for your gene of interest?}
}
\value{
List with trimmed metadata sets for both genes of interest and neighboring genes (additional files generated en route)
}
\description{
Given a desired genome neighborhood and lists of genes of interest and their neighbors, this function removes any genes not on the same scaffold as their ostensible neighbors and trims truncated clusters, removing them from further analysis.
}
\examples{
 neighborTrimOutput <- neighborTrim(imgNeighborsData = imgNeighborsData, imgGenesData = imgGenesData, imgNeighborsContext = imgNeighborsContext, trimShortClusters = TRUE, neighborNumber = 10, geneName = "genE", imgNeighborSeqs=imgNeighborSeqs, imgGeneSeqs=imgGeneSeqs)

}
