% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers2Words.r
\name{numbers2words}
\alias{numbers2words}
\title{Function for turning numbers into words}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{What number do you want turned into a word?}
}
\value{
Lists of neighboring genes to upload to your IMG Gene Cart
}
\description{
This function does what it says on the tin and turns numbers into words.
}
