% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdAlign.r
\name{pcdAlign}
\alias{pcdAlign}
\title{Subfunction to automatically align BGCs for visualization.}
\usage{
pcdAlign(
  finalGeneSets = finalGeneSets,
  standAlone = standAlone,
  uniqueBGCs = uniqueBGCs,
  coreGeneName = coreGeneName
)
}
\arguments{
\item{finalGeneSets}{Partially processed data frame from IMG or analyzeNeighbors with metadata for your gene (family) of interest.  Has already been through the initial prettyClusterDiagrams input process.}

\item{standAlone}{Specifies whether this dataset has been through the rest of the prettyClusters pipeline}

\item{uniqueBGCs}{List of unique BGCs in the dataset}

\item{coreGeneName}{Name of gene (family) of interest - mostly important if this is being run in standAlone mode}
}
\value{
dataframe containing updated metadata with relative positioning information for the final gene cluster diagrams
}
\description{
Aligns BGCs for visualization, centering around a gene (family) of interest.
}
\examples{
\dontrun{
pcdAlign <- pcdAlign(finalGeneSets = finalGeneSets, 
                                   standAlone = standAlone, 
                                   uniqueBGCs = uniqueBGCs,
								  coreGeneName, coreGeneName)
}

}
