% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdAnnotate.r
\name{pcdAnnotate}
\alias{pcdAnnotate}
\title{Subfunction to flag genes for highlighting in prettyClusterDiagrams}
\usage{
pcdAnnotate(
  imgGenes = imgGenes,
  geneSets = geneSets,
  coreGeneList = coreGeneList,
  coreGeneName = coreGeneName,
  annotationGuide = annotationGuide
)
}
\arguments{
\item{imgGenes}{Data frame from IMG or analyzeNeighbors with metadata for your gene (family) of interest.}

\item{geneSets}{Partially processed data frame from IMG or analyzeNeighbors with metadata for your gene (family) of interest.  Has already been through the initial prettyClusterDiagrams input process.}

\item{coreGeneList}{List of genes of interest (your core gene family)}

\item{coreGeneName}{Name assigned to genes of interest.}

\item{annotationGuide}{User-provided table with annotation guidelines.}
}
\value{
Data frame with custom assignments for prettyClusterDiagrams based on protein family criteria
}
\description{
Given IMG-formatted metadata for genomic neighborhoods and a user-provided guide file that defines the families that will be highlighted in the BGC diagrams, this function identifies which genes are hits for these user-provided annotation values.
}
\examples{
\dontrun{
pcdAnnotate <- pcdAnnotate(imgGenes = imgGenes, 
                                   geneSets = geneSets, 
                                   coreGeneList = coreGeneList, 
                                   coreGeneName = coreGeneName, 
                                   annotationGuide = annotationGuide)
}

}
