% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdColor.r
\name{pcdColor}
\alias{pcdColor}
\title{Subfunction to automatically draw up custom palettes for BGCs.}
\usage{
pcdColor(
  colorType = colorType,
  paletteInput = paletteInput,
  processed = processed
)
}
\arguments{
\item{colorType}{R color package from which palettes can be drawn.}

\item{paletteInput}{Specific color scheme within your desired R color package.}

\item{processed}{Partially processed data frame from IMG or analyzeNeighbors with metadata for your gene (family) of interest.  Has already been through the initial prettyClusterDiagrams input process.}
}
\value{
List containing sub-list with custom palette and sub-list with gene-types
}
\description{
Takes user guidelines about R color package and module and the number of gene-types to be highlighted, and designs a palette for the genes.  Adds light/dark colors if over a dozen colors are required, and by default highlights hypothetical proteins ("hyp"), mobile genetic elements ("mge"), and genes not specified for highlighting by the user ("other") with grayscale colors.
}
\examples{
\dontrun{
pcdColor <- pcdColor(colorType = colorType, 
                                   paletteInput = paletteInput, 
                                   processed = processed)
}

}
