% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repnodeTrim.r
\name{repnodeTrim}
\alias{repnodeTrim}
\title{Function for trimming IMG gene metadata and neighbor metadata to match EFI-EST-generated representative nodes}
\usage{
repnodeTrim(
  imgGenes = imgGenes,
  imgNeighbors = imgNeighbors,
  imgGeneSeqs = imgGeneSeqs,
  imgNeighborSeqs = imgNeighborSeqs,
  geneName = geneName,
  efiFullMetadata = efiFullMetadata,
  efiFinalMetadata = efiFinalMetadata
)
}
\arguments{
\item{imgGenes}{Filename for the (edited) IMG metadata file for genes of interest.  Text ("geneFile.txt")}

\item{imgNeighbors}{Filename for the (edited) IMG metadata file for neighboring genes.  Text ("neighborFile.txt")}

\item{imgGeneSeqs}{Filename for the FASTA amino acid sequence file for genes of interest.  Text ("geneFile.fa")}

\item{imgNeighborSeqs}{Filename for the FASTA amino acid sequence file for neighboring genes.  Text ("neighborFile.fa")}

\item{geneName}{Name of gene of interest for filenames. Text ("genE")}

\item{efiFullMetadata}{Filename for the Cytoscape node table for the full EFI-EST network. Text ("fullnetwork.txt")}

\item{efiFinalMetadata}{Filename for the Cytoscape node table for the EFI-EST network at the final repnode \%ID cutoff. Text ("finalnetwork.txt")}
}
\value{
List with the trimmed gene and neighbor metadata
}
\description{
After SSN generation via EFI-EST, this function ties together the EFI-EST gene IDs with gene IDs in existing metadata files and produces trimmed metadata and sequence files for the genes and their neighbors using the list of representative nodes at the desired cutoff.
}
\examples{
repnodeTrimOut <- repnodeTrim(imgGenes = "geneFile.txt", imgNeighbors="neighborFile.txt", imgGeneSeqs = "geneSeqs.fa", imgNeighborSeqs = "neighborSeqs.fa", geneName = "genE", efiFullMetadata = "efiFullMetadata.csv", efiFinalMetadata = "efiFinalMetadata.csv")

}
