% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFasta.r
\name{trimFasta}
\alias{trimFasta}
\title{Trim .fa files based on a list of sequence IDs}
\usage{
trimFasta(seqInput = seqInput, trimInput = trimInput, seqName = seqName)
}
\arguments{
\item{seqInput}{Filename for the FASTA-formatted file containing protein sequences.  Text ("seqInput.fa")}

\item{trimInput}{Filename for the list of desired sequences, as a single column with "gene_oid" as a header.  Text ("trimInput.txt")}

\item{seqName}{Filename identifier to add do your file name, along with the date, trim marker, etc.  Text ("genE_subset")}
}
\value{
Output files and data frame with trimmed sequences.
}
\description{
Simple function for trimming .fa files (based on simplified headers), using a list of sequence IDs.
}
\examples{
\dontrun{
trimFastaOut <- trimFasta(seqInput = "seqInput.fa", 
                          trimInput="trimInput.txt", 
                          seqName = "genE")
}

}
