% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborCatalog.r
\name{neighborCatalog}
\alias{neighborCatalog}
\title{Subfunction to catalog protein families in IMG metadata file}
\usage{
neighborCatalog(
  imgGenesTrimmed = imgGenesTrimmed,
  imgNeighborsTrimmed = imgNeighborsTrimmed,
  geneName = geneName,
  neighborThreshold = neighborThreshold,
  coreGeneName = coreGeneName,
  useInterPro = useInterPro,
  useHypo = useHypo
)
}
\arguments{
\item{imgGenesTrimmed}{Data frame from analyzeNeighbors containing metadata for genes of interest}

\item{imgNeighborsTrimmed}{Data frame from analyzeNeighbors containing metadata for neighbors of genes of interest}

\item{geneName}{Name of gene of interest as string}

\item{neighborThreshold}{Abundance of protein family required to include it in further analysis (0-1)}

\item{coreGeneName}{Name of gene of interest as string, without any suffix.}
}
\value{
List of relevant protein families
}
\description{
This function identifies and quantifies protein families found in genomic neighborhoods (as defined by IMG metadata)
}
\examples{
neighborCatalogOut <- neighborCatalog(imgGenesTrimmed=imgGenesTrimmed, imgNeighborsTrimmed = imgNeighborsTrimmed, neighborNumber=neighborNumber, includeGene=includeGene, geneName=geneName, coreGeneName = coreGeneName)

}
