% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborClusters.r
\name{neighborClusters}
\alias{neighborClusters}
\title{Subfunction to cluster genomic neighborhoods based on the similarity of their protein family sets}
\usage{
neighborClusters(
  imgGenesTrimmed = imgGenesTrimmed,
  imgNeighborsTrimmed = imgNeighborsTrimmed,
  geneName = geneName,
  neighborMatrixData = neighborMatrixData,
  autoClust = autoClust,
  clustMethod = clustMethod,
  alphaVal = alphaVal,
  bootStrap = bootStrap,
  coreGeneName = coreGeneName
)
}
\arguments{
\item{imgGenesTrimmed}{Data frame from analyzeNeighbors containing metadata for genes of interest}

\item{imgNeighborsTrimmed}{Data frame from analyzeNeighbors containing metadata for neighbors of genes of interest}

\item{geneName}{Name of gene of interest as string}

\item{neighborMatrixData}{Matrix from analyzeNeighbors/neighborMatrix containing binary data re: protein family presence in gene clusters}

\item{autoClust}{T/F value determining whether clusters will be automatically assigned}

\item{clustMethod}{String specifying what tool (tidygraph, pvclust) will be used to identify clusters}

\item{alphaVal}{Alpha value cutoff for pvclust (Num, 0-1)}

\item{coreGeneName}{Name of gene of interest as string, without any suffix.}
}
\value{
List with matrix object and updated metadata, with misc. files output along the way
}
\description{
This function cluster-based methods to identify subgroups of gene clusters containing genes of interest that have similar genomic neighborhoods.
}
\examples{
neighborClustersOut <- neighborClusters(imgGenesTrimmed = imgGenesTrimmed, imgNeighborsTrimmed = imgNeighborsTrimmed, geneName = geneName, neighborMatrixData = neighborMatrixData, autoClust = autoClust, clustMethod = clustMethod, alphaVal = alphaVal, bootStrap = bootStrap, coreGeneName = coreGeneName)

}
