% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepNeighbors.r
\name{prepNeighbors}
\alias{prepNeighbors}
\title{Function for preliminary analysis of neighboring genes from IMG}
\usage{
prepNeighbors(
  imgGenes = imgGenes,
  imgNeighbors = imgNeighbors,
  geneSeqs = geneSeqs,
  neighborSeqs = neighborSeqs,
  neighborsContext = neighborsContext,
  geneName = geneName,
  neighborNumber = neighborNumber,
  sysTerm = sysTerm,
  efiRepnodes = FALSE,
  neighborThreshold = 0.025,
  hypoAnalysis = TRUE,
  clustMethod = "tidygraph",
  numThreads = 1,
  alphaVal = 0.95,
  bootStrap = 10,
  pidCutoff = 35,
  trimShortClusters = TRUE
)
}
\arguments{
\item{geneSeqs}{What is the file with the metadata for your gene of interest? Filename as string ("filename.txt")}

\item{neighborSeqs}{What is the file with the metadata for your gene of interest? Filename as string ("filename.txt")}

\item{neighborsContext}{File from neighborGenerate tying genes of interest to their neighbors. Filename as string ("filename.txt")}

\item{geneName}{What is the name of your gene? String ("genE")}

\item{neighborNumber}{How many neighbors do you want to look at on each side of the gene? Integer.}

\item{sysTerm}{What terminal are you using? String value ("wsl", "nix"), no default.}

\item{efiRepnodes}{Does the input dataset consist of EFI repnodes?  T/F, defaults to FALSE.}

\item{neighborThreshold}{What percent of gene clusters should a protein family occur in to be of interest? Number, defaults to 0.05}

\item{hypoAnalysis}{Should hypothetical proteins be clustered and analyzeD? T/F value, defaults to TRUE.}

\item{clustMethod}{What clustering method should be used? String value ("tidygraph", "pvclust"), defaults to "tidygraph".}

\item{numThreads}{How many threads should processes use?  Number depends on your processor, defaults to 1 to be safe.}

\item{alphaVal}{What alpha value cutoff should be used for pvclust? Number from 0-1, defaults to 0.95}

\item{bootStrap}{How many bootstrap rounds does pvclust get? Integer, defaults to 10.}

\item{pidCutoff}{Below what percent ID should edges be deleted? Number from 1-100, defaults to 35.}

\item{trimShortClusters}{Should gene clusters with fewer than the minimum gene number be visualized? T/F value.}

\item{imgGenesTemp}{What is the file with the metadata for your gene of interest? Filename as string ("filename.txt")}

\item{imgNeighborsTemp}{What is the file with the metadata for your gene of interest? Filename as string ("filename.txt")}
}
\value{
List with trimmed metadata sets for both genes of interest and neighboring genes (additional files generated en route)
}
\description{
This function uses IMG metadatasets for genes of interest and their genomic neighbors to do some QC and identify subfamilies of hypothetical proteins.
}
\examples{
prepNeighborsOutput <- prepNeighbors(imgGenes="geneFile.txt", imgNeighbors="neighborsFile.txt", geneSeqs="geneSeqs.fa", neighborSeqs="neighborSeqs.fa", neighborsContext = "context.txt", geneName="genE",  neighborNumber=10, sysTerm="wsl")

}
