% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyClusterDiagrams.r
\name{prettyClusterDiagrams}
\alias{prettyClusterDiagrams}
\title{Umbrella function generates pretty gene cluster diagrams}
\usage{
prettyClusterDiagrams(
  imgGenesFile = imgGenesFile,
  imgNeighborsFile = imgNeighborsFile,
  annotationGuideFile = annotationGuideFile,
  geneName = geneName,
  efiRepnodes = FALSE,
  neighborNumber = neighborNumber,
  annotateGenes = TRUE,
  standAlone = FALSE,
  markClusters = FALSE,
  autoColor = TRUE,
  colorType = "viridis",
  paletteInput = "plasma",
  showScaffold = FALSE,
  alignToCore = TRUE,
  labelGenes = FALSE,
  subclusterDiagrams = FALSE
)
}
\arguments{
\item{imgGenesFile}{What is the file with the metadata for your gene of interest? Filename as string ("filename.txt")}

\item{imgNeighborsFile}{What is the file with the metadata for neighbors of your gene of interest? Filename as string ("filename.txt")}

\item{annotationGuideFile}{User-defined annotation fine, assigning gene names via protein families.  Filename as string ("filename.txt")}

\item{geneName}{What is the name of your gene? Gene name as string ("genE")}

\item{efiRepnodes}{Are values here post-EFI-EST? T/F value, defaults to FALSE}

\item{neighborNumber}{How many genes are expected on either side of the gene of interest? Integer (e.g. 10)}

\item{annotateGenes}{Should the genes be auto-annotated using the geneFormat file? T/F, defaults to TRUE.}

\item{standAlone}{Should we assume the data has not been through the full pipeline? T/F, defaults to FALSE.}

\item{markClusters}{Should cluster labels be included in the final figure? T/F, defaults to FALSE.}

\item{autoColor}{If TRUE, generates palettes automatically. If FALSE, requires hex-formatted colors per gene in the last column of the geneFormat data. Defaults to TRUE.}

\item{colorType}{What palette family should be used? String, defaults to "nord"}

\item{paletteInput}{What specific palette within that family should be used? String, defaults to "aurora"}

\item{showScaffold}{Should the scaffold be included in name labels? T/F, defaults to FALSE}

\item{alignToCore}{Should the core gene be centered in the diagrams? T/F, defaults to TRUE}

\item{labelGenes}{Should all genes be labeled in the diagrams? T/F, defaults to FALSE}

\item{subclusterDiagrams}{Should additional figures be made for each group of gene clusters? T/F, defaults to FALSE.}
}
\value{
Gene diagrams as .png and .pdf files.
}
\description{
This function generates gene cluster diagrams incorporating annotation and clustering info from several sources.
}
\examples{
\dontrun{
prettyClusterOutput <- prettyClusterDiagrams(imgGenesFile = "genesFile.txt", 
                                             imgNeighborsFile = "neighborsFile.txt", 
                                             annotationGuideFile = "formatFile.txt", 
                                             geneName = "genE", 
                                             neighborNumber = 10)
}
 
}
